Refactor endpoint request input helpers
